+++ internal/network/firewall.go
@@ -8,8 +8,9 @@ import (
 )
 
 const (
-	DnsttPort      = "5300"
-	SlipstreamPort = "5301"
+	DnsttPort       = "5300"
+	SlipstreamPort  = "5301"
+	ShadowsocksPort = "5302"
 )
 
 type FirewallType int
@@ -330,14 +331,18 @@ func RemoveAllFirewallRules() {
 	case FirewallFirewalld:
 		removeFirewalldRulesForPort(DnsttPort)
 		removeFirewalldRulesForPort(SlipstreamPort)
+		removeFirewalldRulesForPort(ShadowsocksPort)
 	case FirewallUFW:
 		removeUFWRulesForPort(DnsttPort)
 		removeUFWRulesForPort(SlipstreamPort)
+		removeUFWRulesForPort(ShadowsocksPort)
 	case FirewallIptables, FirewallNone:
 		clearIptablesRulesForPort(DnsttPort)
 		clearIptablesRulesForPort(SlipstreamPort)
+		clearIptablesRulesForPort(ShadowsocksPort)
 		clearIp6tablesRulesForPort(DnsttPort)
 		clearIp6tablesRulesForPort(SlipstreamPort)
+		clearIp6tablesRulesForPort(ShadowsocksPort)
 		saveIptablesRules()
 	}
 }